centralize compose yaml types
