+++ import/compose/parse.go
@@ -1,60 +1,7 @@
-// NOTE [COMPOSE_YAML]: These definitions are spreadout a bit, since the yaml
-// formats are used for component specs.
-//
-// References:
-// https://github.com/compose-spec/compose-spec/blob/master/spec.md
-// https://docs.docker.com/compose/compose-file/compose-file-v3/
-//
-// Fields enumerated as of July 17, 2021 with from the following spec file:
-// <https://github.com/compose-spec/compose-spec/blob/5141aafafa6ea03fcf52eb2b44218408825ab480/spec.md>.
-
 package compose
 
 import (
-	"io"
-
-	"github.com/deref/exo/providers/docker/components/container"
-	"github.com/deref/exo/providers/docker/components/network"
-	"github.com/deref/exo/providers/docker/components/volume"
-	"github.com/goccy/go-yaml"
+	"github.com/deref/exo/providers/docker/compose"
 )
 
-type Compose struct {
-	Version  string             `yaml:"version"`
-	Services map[string]Service `yaml:"services"`
-	Networks map[string]Network `yaml:"networks"`
-	Volumes  map[string]Volume  `yaml:"volumes"`
-	Configs  map[string]Config  `yaml:"configs"`
-	Secrets  map[string]Secret  `yaml:"secrets"`
-	// TODO: extensions with "x-" prefix.
-}
-
-type Service = container.Spec
-
-type Network = network.Spec
-
-type Volume = volume.Spec
-
-type Config struct {
-	File     string `yaml:"file"`
-	External bool   `yaml:"external"`
-	Name     string `yaml:"name"`
-}
-
-type Secret struct {
-	File     string `yaml:"file"`
-	External bool   `yaml:"external"`
-	Name     string `yaml:"name"`
-}
-
-func Parse(r io.Reader) (*Compose, error) {
-	dec := yaml.NewDecoder(r,
-		yaml.DisallowDuplicateKey(),
-		yaml.DisallowUnknownField(), // TODO: Handle this more gracefully.
-	)
-	var comp Compose
-	if err := dec.Decode(&comp); err != nil {
-		return nil, err
-	}
-	return &comp, nil
-}
+var Parse = compose.Parse
