+++ providers/docker/components/network/component.go
@@ -1,6 +1,7 @@
 package network
 
 import (
+	"github.com/deref/exo/providers/docker/compose"
 	docker "github.com/docker/docker/client"
 )
 
@@ -12,17 +13,7 @@ type Network struct {
 	Docker *docker.Client
 }
 
-// See note: [COMPOSE_YAML].
-type Spec struct {
-	Driver     string            `yaml:"driver"`
-	DriverOpts map[string]string `yaml:"driver_opts"`
-	Attachable bool              `yaml:"attachable"`
-	EnableIPv6 bool              `yaml:"enable_ipv6"`
-	Internal   bool              `yaml:"internal"`
-	Labels     map[string]string `yaml:"labels"` // TODO: Support array syntax.
-	External   bool              `yaml:"external"`
-	// TODO: name
-}
+type Spec compose.Network
 
 type State struct {
 	NetworkID string `json:"networkId"`
