+++ providers/docker/components/volume/component.go
@@ -1,6 +1,7 @@
 package volume
 
 import (
+	"github.com/deref/exo/providers/docker/compose"
 	docker "github.com/docker/docker/client"
 )
 
@@ -12,14 +13,7 @@ type Volume struct {
 	Docker *docker.Client
 }
 
-// See note: [COMPOSE_YAML].
-type Spec struct {
-	Driver     string            `yaml:"driver"`
-	DriverOpts map[string]string `yaml:"driver_opts"`
-	// TODO: external
-	Labels map[string]string `yaml:"labels"` // TODO: Support array syntax.
-	Name   string            `yaml:"name"`
-}
+type Spec compose.Volume
 
 type State struct {
 	VolumeName string `json:"volumeId"`
